requirements.txt
azure-storage-blob==2.1.0
python-decouple==3.3
tweepy==3.9.0
